Separación de la libreria en archivos por clase
